+++ src/lib.rs
@@ -1795,13 +1795,12 @@ impl JsError {
     }
 }
 
-#[cfg(feature = "std")]
 impl<E> From<E> for JsError
 where
-    E: std::error::Error,
+    E: core::error::Error,
 {
     fn from(error: E) -> Self {
-        use std::string::ToString;
+        use alloc::string::ToString;
 
         JsError::new(&error.to_string())
     }